+++ Changelog.md
@@ -2,9 +2,13 @@
 
 * Added ability to create reproducible for any step using instance method
   `Step.create_reproducible()`
-* Added ability to load and run Step from Config and State JSON files
+* Added ability to load and run Step from Config and State JSON files, working
+  for both existing step folders and new reproducibles
+* Added new CLI- under either the script `openlane.steps` or
+  `python3 -m openlane.steps`, exposing the two functionalities above.
 * Added internal ability to load a Config without attempting to load the PDK
   configuration data- i.e., only rely on the user's input
+* Extended `Meta` objects to support Step ID and OpenLane version.
 * Moved tests from source tree to `test/` folder, refactoring as necessary
 * Internal `utils` module folded into `common` module, with elements publicly
   documented